Strategy for loading commands.
By simply using a common command prefix, `execute:`, everything becomes
simpler. Filtering of parent commands, if desired, can be done by an
anonymous function in the application. Simply source a file if
necessary, then call the function assuming it's in the file you've
sourced.

We still need an individual file for each command. We can see if we can
live with that. We probably can. Otherwise we can place something on the
first line that we can get with a `head -n 1`, but I don't see why we
need to make it any easier than it already is.