Replace completion fallthrough with delegation.
Instead of falling through, instead invoke a function. We can now name a
function `complete:*` and have that function called during completion.
If no such function exists we instead call `delegate`.

Added the reporting of a message, which we are only going to use for
error messages, for now at least. It works with Bash and will work with
Zsh when we replace direct calls to `compadd` to calls to the completion
system's `_message`.