Correctly update Arch Linux.
We want to upgrade all of Arch Linux at once. We were doing something
where we were updating the repo and then installing a single item, but
it might upgrade a dependency that others depend upon and break the
others. You must update the repo and the system at once. Then isntall
from the update.

    https://gist.github.com/vodik/5660494