Sketch of split args and completion.
Objective is to simplify argument parsing and make it available for use
as a function parser, as it is now, but also split execution from
completion making it easier to descend the commands for completion
without getting intercepted by user commands.

Current rough sketch of completion, seeing first to restore basic
completion for Zsh. Not there yet.