Fix default file completion in Zsh.
The default file completion in Zsh was creating strange output, first
because it did not have a prefix set, subequently the problem was
mysterious, performing the completion but with bad line breaks throwing
everythings.

`zparseopts` called inside the completion function modifies global
completion state even when run in a subshell. This is a known issue with
zparseopts in completion contexts. Replaced it with a call to `args`.