New description system for completions, options.
New description system places the description in the synopsis. It is
removed from the man page display. There are two new pseudo-directives,
.DC to downcase the first character in a sentance, which is only applied
when generating completions, and .JN which is removed so that two
sentances join into the same paragraph, but remain separated for
completion extraction.

Should be stable and not require a major version bump, even if we extend
it to have a different descriptions for inclusion into man pages versus
completions.

Completions will now have their trailing period stripped as well.