Working completion cache.
Uses a two pass system. If an opertion is known to be slow, engaging in
slow network calls, invokeing `zshctl __complete` will return a cache
key and a cache function with arguments. The completion program will
then call `zshctl __encache` with the function and arguments and
evaluate that as output.

It will of course check output in it's cache first.

For a brute invalidation mechanism, we touch a file on the file system
if anything at all cached is updated. Race conditions, so we need to
eventually add a ttl parameter or other ability to clear the completion
cache, but currently operations are unlikely to occur in parallel.

That is, for the administrative tools that use caching, the resources,
like 1Password vaults, are not winking in and out of existance by the
work of a thoursand adminisrators at a thousand typewriters.