Resurrect `block` with optional indent/show.
Made it possible to use tool-native error reporting in ordinary
programs, with some program context for the tool error. By specifying
`block -i 0` the user can skip the indent, work without show. The `try
-v try` pattern is a decent pattern that is appropriately verbose for
what it does. Redirections to variable file handles are not foreign and
should not be avoided.