Use more globs, comment `_zshctl_completions`.
Using globs where we were once using regexen. Allows for the use of
globs directly in the case statement. This is in `_zshctl_help`. Will
continue to follow the pattern of using globs, but if globs won't do,
using `pcre`.

Added comments to `_zshctl_completions`. The intent was difficult to
discern from re-reading. Should be easier now.