Exemplar inline, spaces dedent.
Like `heredoc`, `exemplar` need to dedent when the user has a multi-line
string or command substitution. Just like `heredoc`, when the line
starts with spaces, we count the spaces to the first non-space character
for each line and use the smallest count of spaces as the dedent.

Now able to call exemplar inline, or else in a compiler mode. Exemplar
will only ever compile a function once, so it is safe to call it with
the same function twice.

Fixed a remaining reference to `_zshctl_commands`.